% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann-R.R
\name{diann_matrix}
\alias{diann_matrix}
\title{Generate a matrix with quantities. Zero quantities will be replaced with NA values.}
\usage{
diann_matrix(
  x,
  id.header = "Precursor.Id",
  quantity.header = "Precursor.Normalised",
  proteotypic.only = F,
  q = 0.01,
  protein.q = 1,
  pg.q = 1,
  gg.q = 1
)
}
\arguments{
\item{x}{DIA-NN report}

\item{id.header}{protein/peptide/precursor Id column name}

\item{quantity.header}{quantity column name}

\item{proteotypic.only}{only proteotypic peptides and the respective proteins should be considered?}

\item{q}{precursor q-value threshold}

\item{protein.q}{uniquely identified protein q-value threshold}

\item{pg.q}{protein group q-value threshold}

\item{gg.q}{gene group q-value threshold}
}
\description{
Generate a matrix with quantities. Zero quantities will be replaced with NA values.
}
