% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann-R.R
\name{diann_maxlfq}
\alias{diann_maxlfq}
\title{Quantify using the MaxLFQ algorithm https://doi.org/10.1074/mcp.M113.031591.
This function can be used to calculate protein quantities from peptide/precursor/fragment quantities or, for example, to calculate peptide quantities from precursor/fragment quantities.}
\usage{
diann_maxlfq(
  x,
  sample.header = "File.Name",
  group.header = "Protein.Names",
  id.header = "Precursor.Id",
  quantity.header = "Precursor.Normalised",
  margin = -10
)
}
\arguments{
\item{x}{data frame or data table}

\item{sample.header}{sample Id column name}

\item{group.header}{column name corresponding to the group Id, e.g. protein Id}

\item{id.header}{precursor/peptide Id column name}

\item{quantity.header}{precursor/peptide quantity column name}

\item{margin}{quantities below exp(margin) might be treated as NA}
}
\description{
Quantify using the MaxLFQ algorithm https://doi.org/10.1074/mcp.M113.031591.
This function can be used to calculate protein quantities from peptide/precursor/fragment quantities or, for example, to calculate peptide quantities from precursor/fragment quantities.
}
\examples{
df <- data.frame(File.Name = c("A","A","A","B","B","B"),Protein.Names=rep("ALB",6),
      Precursor.Id=rep(c("PEPTIDE","EPTIDEP","PTIDEPE"),2),Precursor.Normalised=c(20,10,5,25,12,NA))
diann_maxlfq(df)
}
