% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann-R.R
\name{diann_save}
\alias{diann_save}
\title{Save a data frame to a tab-separated file}
\usage{
diann_save(x, file, row.names = F)
}
\arguments{
\item{x}{data frame or data table}

\item{file}{file name}

\item{row.names}{save row names?}
}
\description{
Save a data frame to a tab-separated file
}
